models/ai_analyzer.py
from sklearn.ensemble import RandomForestClassifier
import joblib, numpy as np

class AIAnalyzer:
    def __init__(self):
        self.model = RandomForestClassifier(n_estimators=100)
        self.train_model()
    
    def analyze(self, vulns):
        # Feature extraction: port, service age, known CVEs
        features = np.array([[len(v['name']), v['cvss'], 1] for v in vulns])
        risks = self.model.predict_proba(features)[:, 1] * 100
        return [{'risk_score': r, 'recommendation': self.get_fix(v)} for v, r in zip(vulns, risks)]