requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
reportlab==4.0.4
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
matplotlib==3.7.2
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
PyMySQL==1.1.0
requests==2.31.0
python-nmap==0.7.1
OWASP-ZAP==2.14.0
flask==2.3.3
python-nmap==0.7.1
reportlab==4.0.4
scikit-learn==1.3.2
gunicorn==21.2.0